requirements.txt
llama-index
llama-index-graph-stores-neo4j
llama-index-llms-gemini
llama-index-embeddings-google
neo4j
google-generativeai
unstructured[pdf]
pymupdf
python-dotenv
pandas
ragas
deepeval
